Task # 13 Configured app with auto-ddl=validate and flyway enabled

+++ src/main/java/pl/futurecollars/invoicing/db/Database.java
@@ -10,15 +10,15 @@ import pl.futurecollars.invoicing.model.InvoiceEntry;
 
 public interface Database {
 
-    int save(Invoice invoice);
+    long save(Invoice invoice);
 
-    Optional<Invoice> getById(int id);
+    Optional<Invoice> getById(long id);
 
     List<Invoice> getAll();
 
-    Optional<Invoice> update(int id, Invoice updatedInvoice);
+    Optional<Invoice> update(long id, Invoice updatedInvoice);
 
-    Optional<Invoice> delete(int id);
+    Optional<Invoice> delete(long id);
 
     default BigDecimal visit(
         Predicate<Invoice> invoicePredicate,
